+++ fulltext/backends/__xlsx.py
@@ -1,7 +1,7 @@
-from __future__ import absolute_import
+from io import StringIO
+import datetime
 
-import xlrd
-from six import StringIO
+import openpyxl
 
 from fulltext.util import BaseBackend
 from fulltext.util import assert_cmd_exists, exiftool_title
@@ -13,21 +13,26 @@ class Backend(BaseBackend):
         if title:
             assert_cmd_exists('exiftool')
 
-    def handle_path(self, path):
+    def handle_fobj(self, fobj):
         text = StringIO()
-        wb = xlrd.open_workbook(path)
-        for n in wb.sheet_names():
-            ws = wb.sheet_by_name(n)
-            for x in range(ws.nrows):
-                for y in range(ws.ncols):
-                    v = ws.cell_value(x, y)
+        wb = openpyxl.load_workbook(fobj)
+        for n in wb.sheetnames:
+            ws = wb[n]
+            for x in range(1, ws.max_row+1):
+                for y in range(1, ws.max_column+1):
+                    v = ws.cell(x, y).value
                     if v:
-                        if isinstance(v, (int, float)):
+                        if isinstance(v, (int, float, datetime.datetime)):
                             v = str(v)
                         text.write(v)
-                        text.write(u' ')
+                        text.write(u'\t')
                 text.write(u'\n')
         return text.getvalue()
+    
+    def handle_path(self, path):
+        with open(path, 'rb') as fin:
+            return self.handle_fobj(fin)
 
     def handle_title(self, f):
         return exiftool_title(f, self.encoding, self.encoding_errors)
+
