+++ requirements.txt
@@ -1,20 +1,18 @@
-xlrd
-bs4
-docx2txt
-lxml
-python-pptx
-six
-docopt
-pytesseract
-Pillow
-pycodestyle
-pyflakes
-python-magic
+xlrd~=2.0.1
+beautifulsoup4~=4.9.3
+docx2txt~=0.8
+lxml~=4.6.3
+python-pptx~=0.6.19
+six~=1.16.0
+docopt~=0.6.2
+pytesseract~=0.3.8
+Pillow~=8.3.1
+python-magic~=0.4.24
 ebooklib
-contextlib2
-mock
-rarfile
-flake8
+contextlib2~=21.6.0
+mock~=4.0.3
+rarfile~=4.0
+openpyxl~=3.0.7
 
 git+https://github.com/mattgwwalker/msg-extractor.git@2a24c9950d34932ed5979693cb70d758d78715df#egg=ExtractMsg
